refactor(kernel): extract advisory-lock syscalls to kernel/locks.rs
Phase G.2 of Phase 3 restructure plan.  First per-syscall-family
extraction.

  - rust/kernel/src/kernel/locks.rs (new): impl Kernel block over
    sys_lock / sys_unlock / metastore_list_locks /
    install_federation_locks.  ~95 LOC.
  - rust/kernel/src/kernel/mod.rs: drops the corresponding methods,
    leaves a `// (Moved to kernel::locks submodule)` breadcrumb at
    the section divider, declares `mod locks;` in the new
    "Phase G: per-syscall-family submodules" header.

No semantic change — each method stays a member of Kernel via the
submodule's `impl Kernel { ... }` block.

cargo check -p kernel: clean.
